# antimoodlo-api/Handlers/AnswerHandler.go
package Handlers

import (
	"antimoodlo/Models"
	"antimoodlo/db"
	"github.com/gin-gonic/gin"
	"net/http"
)

// AddQuestionOption godoc
// @Summary      Добавить вариант ответа
// @Description  Добавляет один из вариантов ответа для вопроса (Multiple/Single Choice)
// @Tags         answers
// @Accept       json
// @Produce      json
// @Param        id      path      int                     true  "ID вопроса"
// @Param        option  body      Models.QuestionOption   true  "Вариант ответа"
// @Success      201     {object}  Models.QuestionOption
// @Failure      400     {object}  map[string]interface{}
// @Failure      404     {object}  map[string]interface{}
// @Failure      500     {object}  map[string]interface{}
// @Router       /questions/{id}/options [post]
func AddQuestionOption(c *gin.Context) {
	var option Models.QuestionOption
	questionID := parseUint(c.Param("id"))

	if err := DB.DB.First(&Models.Question{}, questionID).Error; err != nil {
		c.JSON(http.StatusNotFound, gin.H{"error": "Question not found"})
		return
	}

	if err := c.ShouldBindJSON(&option); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	option.QuestionID = questionID
	if err := DB.DB.Create(&option).Error; err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to create option"})
		return
	}

	c.JSON(http.StatusCreated, option)
}

// AddCorrectAnswer godoc
// @Summary      Добавить правильный ответ
// @Description  Добавляет правильный вариант ответа к вопросу (по ID)
// @Tags         answers
// @Accept       json
// @Produce      json
// @Param        id      path      int                    true  "ID вопроса"
// @Param        answer  body      Models.CorrectAnswer   true  "Правильный ответ"
// @Success      201     {object}  Models.CorrectAnswer
// @Failure      400     {object}  map[string]interface{}
// @Failure      404     {object}  map[string]interface{}
// @Failure      500     {object}  map[string]interface{}
// @Router       /questions/{id}/answers/correct [post]
func AddCorrectAnswer(c *gin.Context) {
	var answer Models.CorrectAnswer
	questionID := parseUint(c.Param("id"))

	if err := DB.DB.First(&Models.Question{}, questionID).Error; err != nil {
		c.JSON(http.StatusNotFound, gin.H{"error": "Question not found"})
		return
	}

	if err := c.ShouldBindJSON(&answer); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	answer.QuestionID = questionID
	if err := DB.DB.Create(&answer).Error; err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to create correct answer"})
		return
	}

	c.JSON(http.StatusCreated, answer)
}

// AddOpenAnswer godoc
// @Summary      Добавить открытый ответ
// @Description  Добавляет открытый ответ для вопроса с открытой формой
// @Tags         answers
// @Accept       json
// @Produce      json
// @Param        id      path      int                 true  "ID вопроса"
// @Param        answer  body      Models.OpenAnswer   true  "Открытый ответ"
// @Success      201     {object}  Models.OpenAnswer
// @Failure      400     {object}  map[string]interface{}
// @Failure      404     {object}  map[string]interface{}
// @Failure      500     {object}  map[string]interface{}
// @Router       /questions/{id}/answers/open [post]
func AddOpenAnswer(c *gin.Context) {
	var answer Models.OpenAnswer
	questionID := parseUint(c.Param("id"))

	if err := DB.DB.First(&Models.Question{}, questionID).Error; err != nil {
		c.JSON(http.StatusNotFound, gin.H{"error": "Question not found"})
		return
	}

	if err := c.ShouldBindJSON(&answer); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	answer.QuestionID = questionID
	if err := DB.DB.Create(&answer).Error; err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to create open answer"})
		return
	}

	c.JSON(http.StatusCreated, answer)
}

// AddMatchPair godoc
// @Summary      Добавить пару для сопоставления
// @Description  Добавляет пару "лево-право" для вопросов типа Match
// @Tags         answers
// @Accept       json
// @Produce      json
// @Param        id    path      int              true  "ID вопроса"
// @Param        pair  body      Models.MatchPair  true  "Пара для сопоставления"
// @Success      201   {object}  Models.MatchPair
// @Failure      400   {object}  map[string]interface{}
// @Failure      404   {object}  map[string]interface{}
// @Failure      500   {object}  map[string]interface{}
// @Router       /questions/{id}/answers/match [post]
func AddMatchPair(c *gin.Context) {
	var pair Models.MatchPair
	questionID := parseUint(c.Param("id"))

	if err := DB.DB.First(&Models.Question{}, questionID).Error; err != nil {
		c.JSON(http.StatusNotFound, gin.H{"error": "Question not found"})
		return
	}

	if err := c.ShouldBindJSON(&pair); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	pair.QuestionID = questionID
	if err := DB.DB.Create(&pair).Error; err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to create match pair"})
		return
	}

	c.JSON(http.StatusCreated, pair)
}